requirements.txt
PyYAML
youtube_dl
